tests/eas: Un-factorise code that is no longer shared
tests/eas/__init__.py used to contain code shared between generic.py and
acceptance.py. Since the latter has been removed, we can move that functionality
into the former and delete __init__.py
